+++ src/main/scala/com/urbancomputing/just/kvsearch/util/DistanceUtils.scala
@@ -6,15 +6,15 @@ object DistanceUtils {
   /**
    * euclidean distance
    */
-  def euclideanDistance(s1: Seq[Double], s2: Seq[Double]): Double = {
+  def euclideanDistance(s1: Seq[Float], s2: Seq[Float]): Float = {
     assert(s1.length == s2.length)
-    sqrt(s1.zip(s2).map(p => pow(p._1 - p._2, 2)).sum)
+    sqrt(s1.zip(s2).map(p => pow(p._1 - p._2, 2)).sum).toFloat
   }
 
   /**
    * manhattan distance
    */
-  def manhattanDistance(s1: Seq[Double], s2: Seq[Double]): Double = {
+  def manhattanDistance(s1: Seq[Float], s2: Seq[Float]): Float = {
     assert(s1.length == s2.length)
     s1.zip(s2).map(p => abs(p._1 - p._2)).sum
   }
@@ -22,7 +22,7 @@ object DistanceUtils {
   /**
    * longest common sub sequence
    */
-  def longestCommonSubSeq(s1: Seq[Double], s2: Seq[Double], threshold: Double): Int = {
+  def longestCommonSubSeq(s1: Seq[Float], s2: Seq[Float], threshold: Float): Int = {
     assert(s1.length == s2.length)
     s1.zip(s2).map(p => abs(p._1 - p._2) <= threshold).count(_ == true)
   }
@@ -30,7 +30,7 @@ object DistanceUtils {
   /**
    * chebyshev distance
    */
-  def chebyshevDistance(s1: Seq[Double], s2: Seq[Double]): Double = {
+  def chebyshevDistance(s1: Seq[Float], s2: Seq[Float]): Float = {
     assert(s1.length == s2.length)
     s1.zip(s2).map(p => abs(p._1 - p._2)).max
   }
@@ -38,7 +38,7 @@ object DistanceUtils {
   /**
    * block distance: lower bound
    */
-  def blockDistance(a: (Double, Double), b: (Double, Double)): Double = {
+  def blockDistance(a: (Float, Float), b: (Float, Float)): Float = {
     val (aMax, aMin) = a
     val (bMax, bMin) = b
     Seq(
@@ -52,7 +52,7 @@ object DistanceUtils {
   /**
    * multi block distance: lower bound
    */
-  def multiBlockDistance(m1: Seq[(Double, Double)], m2: Seq[(Double, Double)]): Double = {
+  def multiBlockDistance(m1: Seq[(Float, Float)], m2: Seq[(Float, Float)]): Float = {
     assert(m1.length == m2.length)
     m1.zip(m2).map(b => blockDistance(b._1, b._2)).max
   }
