+++ src/main/scala/com/urbancomputing/just/kvsearch/util/OtherUtils.scala
@@ -25,7 +25,7 @@ object OtherUtils {
   /**
    * print result in short
    */
-  def printRes(res: Seq[(Int, Double)]): String = {
+  def printRes(res: Seq[(Int, Float)]): String = {
     val resNum = res.length
     resNum match {
       case 0 => "null"
@@ -41,28 +41,28 @@ object OtherUtils {
    * @param line id \t v1,v2,v3,...
    * @return (id, seq)
    */
-  def splitLine(line: String): (Int, Seq[Double]) = {
+  def splitLine(line: String): (Int, Seq[Float]) = {
     val idAndSeq = line.split("\t", 2)
     val id = idAndSeq.head.toInt
-    val seq = idAndSeq.last.split(",").map(_.toDouble)
+    val seq = idAndSeq.last.split(",").map(_.toFloat)
     (id, seq)
   }
 
   /**
    * split input query seq
    */
-  def splitQs(qs: String, sep1: String = "#", sep2: String = ","): Seq[(Int, Double)] = {
+  def splitQs(qs: String, sep1: String = "#", sep2: String = ","): Seq[(Int, Float)] = {
     qs.split(sep1)
-      .map(t => (t.split(sep2).head.toInt, t.split(",").last.toDouble))
+      .map(t => (t.split(sep2).head.toInt, t.split(",").last.toFloat))
   }
 
   /**
    * generate gaussian between 0 and 1
    */
-  def genGaussianZeroToOne(): Double = {
-    var res: Double = Double.MaxValue
+  def genGaussianZeroToOne(): Float = {
+    var res: Float = Float.MaxValue
     while (res < 0.0 || res > 1.0) {
-      res = Random.nextGaussian()
+      res = Random.nextGaussian().toFloat
     }
     res
   }
@@ -72,11 +72,11 @@ object OtherUtils {
    *
    * @todo implement with RW or OU
    */
-  def genOneFakeSeq(seq: Seq[Double], fluctuateRate: Double = 0.1): Seq[Double] = {
+  def genOneFakeSeq(seq: Seq[Float], fluctuateRate: Float = 0.1F): Seq[Float] = {
     val (seqMax, seqMin) = (seq.max, seq.min)
     seq.map(v => {
       val fakeValue = max(v + (OtherUtils.genGaussianZeroToOne() - 0.5) * ((seqMax - seqMin) / 2) * fluctuateRate, 0.0)
-      fakeValue.formatted("%.5f").toDouble
+      fakeValue.formatted("%.5f").toFloat
     })
   }
 
@@ -85,8 +85,8 @@ object OtherUtils {
    *
    * @return [(time block index, seq of max-min of time value block, seq of value)]
    */
-  def genTuple3(seq: Seq[(Int, Double)], timeBlockLen: Int,
-                valueBlockLen: Int): Seq[(Int, Seq[(Int, Double, Double)], Seq[(Int, Double)])] = {
+  def genTuple3(seq: Seq[(Int, Float)], timeBlockLen: Int,
+                valueBlockLen: Int): Seq[(Int, Seq[(Int, Float, Float)], Seq[(Int, Float)])] = {
     seq
       .map(t => (t._1 / timeBlockLen, t._1 % timeBlockLen, t._2))
       .groupBy(_._1)
